backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.12.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
uvicorn==0.27.1
Werkzeug==3.0.1
